+++ telegram/__init__.py
@@ -3,3 +3,18 @@ This is just a special file that tells pip that your main module is in this fold
 No need to add anything here. Feel free to delete this line when you make your own package
 Leave it empty
 """
+# flake8: noqa
+from .constants import DEFAULT_PARSE_MODE, DEFAULT_TIMEOUT
+from .telegram import TelegramBot, ExceptionHandler
+from .exceptions import VaultInstanceNotSet, BotNameNotSet, InvalidTokenConfiguration, FailedToCreateInstance
+
+__all__ = [
+    "TelegramBot",
+    "ExceptionHandler",
+    "DEFAULT_PARSE_MODE",
+    "DEFAULT_TIMEOUT",
+    "VaultInstanceNotSet",
+    "BotNameNotSet",
+    "InvalidTokenConfiguration",
+    "FailedToCreateInstance"
+]