+++ telegram/constants.py
@@ -1,0 +1,6 @@
+"""This module contains constants for the telegram package."""
+# Constants
+# The default parse mode
+DEFAULT_PARSE_MODE = 'HTML'
+# The default timeouts when errors occur
+DEFAULT_TIMEOUT = 60